backend/requirements.txt
fastapi
uvicorn[standard]
python-multipart
opencv-python
Pillow
pdf2image
tensorflow
numpy
python-jose[cryptography]